Authen API

+++ ndd-app/booking/views/booking_delete_view.py
@@ -11,12 +11,14 @@ from .booking_page_view import api_filter_bookings
 
 @csrf_exempt
 def api_delete_bookings(request):
-    if request.method == "POST":
-        req = json.loads( request.body.decode('utf-8') )
-        pk_list = req["checked_bookings"]
+    if request.user.is_authenticated:
+        if request.method == "POST":
+            req = json.loads( request.body.decode('utf-8') )
+            pk_list = req["checked_bookings"]
 
-        for pk in pk_list:
-            booking = Booking.objects.get(pk=pk)
-            booking.delete()
+            for pk in pk_list:
+                booking = Booking.objects.get(pk=pk)
+                booking.delete()
 
-    return api_filter_bookings(request)
+        return api_filter_bookings(request)
+    return JsonResponse('Error', safe=False)
